+++ src/jhson.js
@@ -150,11 +150,11 @@
                 childElementData = getElementObject( child, properties, getParentCssStylesCopy( parentCssStyles ) ),
                 addChild = false;
 
-            if ( _configuration.formattingNodeTypes.indexOf( childElementData.nodeName.toLowerCase() ) > _value.notFound ) {
+            if ( _configuration.formattingNodeTypes.indexOf( childElementData.nodeName ) > _value.notFound ) {
                 totalChildren++;
             } else {
 
-                if ( properties.ignoreNodeTypes.indexOf( childElementData.nodeName.toLowerCase() ) === _value.notFound ) {
+                if ( properties.ignoreNodeTypes.indexOf( childElementData.nodeName ) === _value.notFound ) {
                     addChild = true;
                     totalChildren++;
                 }