+++ docs/CHANGE_LOG.md
@@ -4,6 +4,7 @@
 
 #### **Fixes & Improvements:**
 - Fixed a minor formatting issue in the "package.json" file.
+- Updated README.md and README_NUGET.md to include the latest chained function calls.
 
 <br>