+++ docs/CHANGE_LOG.md
@@ -4,6 +4,7 @@
 
 #### **Public Functions:**
 - When "addCssToHead()" is used for "html()", all the CSS is now placed in one "style" tag in the HTML document head, instead of multiple ones!
+- When "addCssToHead()" is used for "html()", the CSS will now use the CSS class from the DOM element (it reverts back to the DOM element ID if not available).
 
 #### **Fixes & Improvements:**
 - Fixed a minor formatting issue in the "package.json" file.