+++ docs/CHANGE_LOG.md
@@ -5,7 +5,7 @@
 #### **Public Functions:**
 - BREAKING: Renamed "json() > includeCssStyles()" to "json() > includeCssProperties()".
 - When "addCssToHead()" is used for "html()", all the CSS is now placed in one "style" node in the HTML document head, instead of multiple ones!
-- When "addCssToHead()" is used for "html()", the CSS will now use the CSS class from the DOM element (it reverts back to the DOM element ID if not available).
+- When "addCssToHead()" is used for "html()", the CSS will now use the CSS class from the DOM element (it reverts to the DOM element ID if not available).
 - Added "ignoreNodeTypes()" support for "json()", which states the node types that should not be included in the JSON returned from "get()".
 - Added "ignoreCssProperties()" support for "json()", which states the CSS properties that should not be included in the JSON returned from "get()".
 - Added "ignoreAttributes()" support for "json()", which states the attributes that should not be included in the JSON returned from "get()".
@@ -18,7 +18,7 @@
 - Fixed a minor formatting issue in the "package.json" file.
 - Fixed a fault that caused the text to be ignored for some elements.
 - Fixed HTML testing files not referencing some of the recent changes.
-- Fixed issues with lowercase usages being done multiple times.
+- Fixed issues with lowercase usage being done multiple times.
 - Updated README.md and README_NUGET.md to include the latest chained function calls.
 - Updated README.md and README_NUGET.md to include the new features added recently.